services/influencer_crm/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
sqlalchemy>=2.0
pydantic>=2.6
psycopg2-binary>=2.9
python-dotenv>=1.0
PyJWT>=2.8
